fix: restructuring the tests folderts
