+++ libs/utils/exportFile.ts
@@ -1,0 +1,7 @@
+import fs from "fs";
+
+const exportFile = async (pathFileName: string, content: string) => {
+  fs.writeFileSync(pathFileName, content);
+};
+
+export default exportFile;
